FIX: compile error (fixes #1687)
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@14405 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ src/Preferences/TMSPreferencesDialog.cpp
@@ -34,7 +34,7 @@ void TMSPreferencesDialog::addServer(const TmsServer & srv)
 	theTmsServers.push_back(srv);
 	if (!srv.deleted) {
 		QListWidgetItem* item = new QListWidgetItem(srv.TmsName);
-		item->setData(Qt::UserRole, theTmsServers.size()-1);
+		item->setData(Qt::UserRole, (int) theTmsServers.size()-1);
 		lvTmsServers->addItem(item);
 	}
 }
